refactor(harness): extract artifact values and recipes
