+++ src/aec_bench/harness/artifact/__init__.py
@@ -1,0 +1,2 @@
+# ABOUTME: Owns effect-free artifact attempt values and recipe policies.
+# ABOUTME: Keeps local task execution orchestration separate from its small value contracts.
